Add a message of whether or not an update is available.
Make a request to extract thelatest version of atmosphere on github, and
display it to the user if an update is available. If the user is up to
date, show "Up to date."

Example up-to-date: https://i.imgur.com/O6MGc8A.jpg
Example update available: https://i.imgur.com/AphtmC7.jpg

NOTE: The application is single threaded, so to give better perceived
latency, I updated the code to show a partially loaded splash screen:
https://i.imgur.com/KR1psQA.jpg. Ideally this should be fetched in the
background.